Remove 'Maker' classes & move helpers to class 'Component'
