--- knv_cli/structure/maker.py
@@ -1,9 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Maker(ABC):
-    # CORE methods
-
-    @abstractmethod
-    def build(self) -> None:
-        pass
